bench: Introduce new progress report and intervals
This commit renews the progress report that bench has by allowing to
report bandwidth results during benchmark.

Also, an additional functionality is that of report intervals. Report
intervals allow the user to choose exactly when progress report will be
printed.

Both of these features are documented extensively in the usage section.