segdev: Fix segment mapping to user.
Do not use remap_pfn_range anymore. This causes problems when a peer performs
direct IO with buffer on the segment. Use tha fault vm operation to dynamically
map a page to the userspace peer address space when it tries to access a
specific part of the segment for the first time.