mapper: Do not drop cache on close explicitly.
Do not drop on cache on close operation explicitly. Instead, when closing a map,
wait all pending I/O operations and then close map. If this succceeds, the map
is marked as non-Exclusive. The map action codepath will always drop the cache
when a map is marked as non Exclusive.