mapper: Change order when writing map.
When writing a map, change the order in which the map metadata and map data are
written. The map metadata is a small block of bytes while the map data can
consist of larger blocks. By doing so, and by always reading the map in the
reverse order (that is first the metadata and then the data), we can create the
illusion of atomic map creation. A map where the metadata have not been written
yet, it can not be loaded and thus cannot get partial data.