archipelago: Pause/Unpause tapdisk on user restart
When archipelago is invoked with '-u' argument tapdisk is paused or unpaused
accordingly. On stop all tapdisk processes, if any, are paused, peer's are
stopped and the shared memory segment is destroyed.
On start the shared memory segment is created, all peer's are started and
finally all tapdisk processes, if any, are unpaused.