mapper: Introduce map signature
Introduce map signature to identify archipelago maps. The signature is 4 bytes
long and represents the string "AMF."

If the signature is not matched, then if the mapper identifies the mapfile as
version 1, it proceeds. Otherwise, we rely upon a user provided assume_v0 flag
to interpret the mapfile as a Pithos (version 0) mapfile. Moreover, mapper now
implicitly and automatically upgrades the version 0 mapfiles to the latest
version.

Also, mapper does not implicitly creates separate namespaces for archipelago
and Pithos maps (version 0) by default. So the "archip_" prefix is not inserted
for objects that represent  map data or volume data.