mapper: Add support for map renaming.
Support map renaming functionality. It is currently implemented as the
equivalent of a copy and a delete functionality.

The current implementations is problematic, if the user attempts to rename a
mapped volume from the node on which it is mapped. In that case, the rename will
proceed and the mapped volume will experience I/O errors.

The solution to the above problem is not trivial and will be addressed in future
commits.