archipelago: Introduce BLKTAP_ENABLED conf variable
Set BLKTAP_ENABLED to True or False to control whether the
'archipelago' and 'vlmc' commands should check for the
existence of the blktap module.

If BLKTAP_ENABLED is set to False, the 'vlmc' command will not
interact with blktap and it won't unload the blktap module if already
loaded. This gives the administrator the ability to continue working
with blktap independently, without interfering with Archipelago.

The intention for this addition is to support Archipelago deployments
that do not need to have peers that interact with blktap.