archipelago: Introduce BLKTAP_ENABLED conf variable
Set True or False BLKTAP_ENABLED configuration variable to control
whether 'archipelago' and 'vlmc' commands should check for the
existence of the blktap module.
If blktap module is disabled 'vlmc' command won't proceed with
it's operations and it won't unload blktap module if already
loaded. This gives the ability to the administrator to continue
working with blktap independently without interfering with Archipelago.
The intention for this addition is to support peers that use Archipelago
without the need of blktap.