mapper: Use fallback map with the 'archip_' prefix
If loading a map with its name fails, try falling back to the volume prefixed by
'archip_'. Then, if this succeeds and we have exclusive access to the map, to a
permanent rename to get rid of the archip prefix.

The above is needed to fasciliate live volume migrations to the new naming
scheme.