filed: Add support for Pithos object migration
Pithos places its objects under the hexlified hash of the contents. It also
creates the dir hierarchy according to the first 6 characters of the object
name. In contrast, filed hashes the object name to get 6 random characters for
the dir hierarchy. In this commit, we introduce a fallback path for filed to
find objects created by Pithos and move them to the directory expected by filed.