blktap: Handle various tapdisk states
Until now the paused state was hardcoded as the '0x2a' value
which corresponds only to paused state that hasn't dropped the
logfile. This commit adds all possible tapdisk states and
uses those states to find if the block device is in paused state,
unpaused state or in pausing state.